requirements.txt
numpy
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary
python-dotenv